requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pdfplumber==0.10.3
python-docx==1.0.1
python-magic-bin==0.4.14
pydantic==2.7.1
typer==0.9.0
rich==13.7.1
pytest==8.1.1
black==24.3.0
ruff==0.4.3
openai==1.30.0  # optional, used only if OPENAI_API_KEY present
flask==3.0.3
flask-cors==4.0.0
